+++ fastly/block_fastly_service_logging_papertrail.go
@@ -8,7 +8,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 
-	gofastly "github.com/fastly/go-fastly/v11/fastly"
+	gofastly "github.com/fastly/go-fastly/v12/fastly"
 )
 
 // PaperTrailServiceAttributeHandler provides a base implementation for ServiceAttributeDefinition.