+++ fastly/block_fastly_service_settings.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 
-	gofastly "github.com/fastly/go-fastly/v11/fastly"
+	gofastly "github.com/fastly/go-fastly/v12/fastly"
 )
 
 // SettingsServiceAttributeHandler provides a base implementation for ServiceAttributeDefinition.