+++ fastly/resource_fastly_ngwaf_account_signal_test.go
@@ -9,8 +9,8 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 
-	"github.com/fastly/go-fastly/v11/fastly/ngwaf/v1/common"
-	"github.com/fastly/go-fastly/v11/fastly/ngwaf/v1/signals"
+	"github.com/fastly/go-fastly/v12/fastly/ngwaf/v1/scope"
+	"github.com/fastly/go-fastly/v12/fastly/ngwaf/v1/signals"
 )
 
 func TestAccFastlyNGWAFAccountSignal_basic(t *testing.T) {
@@ -70,8 +70,8 @@ func testAccCheckNGWAFAccountSignalDestroy(s *terraform.State) error {
 
 		_, err := signals.Get(context.TODO(), conn, &signals.GetInput{
 			SignalID: &rs.Primary.ID,
-			Scope: &common.Scope{
-				Type:      common.ScopeTypeAccount,
+			Scope: &scope.Scope{
+				Type:      scope.ScopeTypeAccount,
 				AppliesTo: []string{"*"},
 			},
 		})