+++ fastly/resource_fastly_ngwaf_alert_pagerduty_integration.go
@@ -10,8 +10,8 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 
-	gofastly "github.com/fastly/go-fastly/v11/fastly"
-	pagerdutyAlerts "github.com/fastly/go-fastly/v11/fastly/ngwaf/v1/workspaces/alerts/pagerduty"
+	gofastly "github.com/fastly/go-fastly/v12/fastly"
+	pagerdutyAlerts "github.com/fastly/go-fastly/v12/fastly/ngwaf/v1/workspaces/alerts/pagerduty"
 )
 
 func resourceFastlyNGWAFAlertPagerDutyIntegration() *schema.Resource {