+++ fastly/resource_fastly_ngwaf_workspace_signal_test.go
@@ -12,8 +12,8 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/fastly/go-fastly/v11/fastly/ngwaf/v1/common"
-	"github.com/fastly/go-fastly/v11/fastly/ngwaf/v1/signals"
+	"github.com/fastly/go-fastly/v12/fastly/ngwaf/v1/scope"
+	"github.com/fastly/go-fastly/v12/fastly/ngwaf/v1/signals"
 )
 
 func TestFlattenNGWAFSignalResponse(t *testing.T) {
@@ -97,8 +97,8 @@ func testAccCheckNGWAFWorkspaceSignalDestroy(s *terraform.State) error {
 
 		_, err := signals.Get(context.TODO(), conn, &signals.GetInput{
 			SignalID: &rs.Primary.ID,
-			Scope: &common.Scope{
-				Type:      common.ScopeTypeWorkspace,
+			Scope: &scope.Scope{
+				Type:      scope.ScopeTypeWorkspace,
 				AppliesTo: []string{wsID},
 			},
 		})